MB-22082: Log if cannot find connection when requesting takeover stats
The referenced MB is failing rebalance because when dcp takeover stats
are requested the specified connection cannot be found.

Whilst the root cause of the MB remains outstanding, it is useful for
debugging purposes to log at the NOTICE level when we cannot find a
connection whilst requesting takeover stats.

Change-Id: I7ff432642fe9598c9f4ef6d64c658d38c610002c
Reviewed-on: http://review.couchbase.org/71643
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>