MB-22241: Only check opaque validity for stream request responses
The response handler attempts to find the opaque in the opaque map for
all responses. However it should only do this for stream request
responses.

If it could not find the opaque in the map it prints a message stating
"Received response with opaque [x] and that stream no longer exists".
See below for an example, where the responses related to control
messages.

2017-01-10T07:20:42.972817Z NOTICE (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - (vb 1023) Attempting to add  stream with start seqno 0, end seqno 18446744073709551615, vbucket uuid 107579528815810, snap start seqno 0, snap end seqno 0, and vb_high_seqno 0
2017-01-10T07:20:42.974227Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 2 and that stream no longer exists
2017-01-10T07:20:42.975006Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 3 and that stream no longer exists
2017-01-10T07:20:42.975481Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 4 and that stream no longer exists
2017-01-10T07:20:42.976090Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 5 and that stream no longer exists
2017-01-10T07:20:42.976710Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 6 and that stream no longer exists
2017-01-10T07:20:42.976943Z WARNING (default) DCP (Consumer) eq_dcpq:replication:n_0@192.168.1.81->n_1@127.0.0.1:default - Received response with opaque 7 and that stream no longer exists

This patch changes the code to only attempt to find the opaque in the
opaque map for stream request responses.  For stream request responses
if an entry cannot be found in the map then a WARNING message is
logged.

If an opaque is found then a subsequent check is made to see if the
opaque held in the map is valid.  If it is not valid a different
WARNING message is logged.

Change-Id: Iee23a7ac0d3dd8455e0462b7b851be3f78601e72
Reviewed-on: http://review.couchbase.org/71762
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>