CheckpointTest: Convert to Typed Test
To expand the Checkpoint Test so it can be run using either a
(persistent) VBucket or EphemeralVBucket object, refactor the
CheckpointTest to be parameterized by type.

Initially it is only Instantiated once with VBucket (as
previously). Once the EphemeralVBucket is introduced it can be added
to the set of instantiations.

Change-Id: I20b2d3ea08c8df0d415e55f580011fc020720e7a
Reviewed-on: http://review.couchbase.org/71379
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>