Reapply "Preserve XATTRs with mutation commands"
This commit brings back the preserve XATTRs with mutation commands.
It has been modified from the original patch by checking if
xattrs is enabled before trying to preserve the xattrs. The reason
for this change is that preserving XATTRs is a "performance regression"
as each mutation needs to fetch the previous value first in order
to copy them over.

Change-Id: I2470e96618f8034f29605382513440fb3857792e
Reviewed-on: http://review.couchbase.org/71860
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>