[Refactor]: Remove unnecessary queueDirty call from completeBGFetchForSingleItem
In completeBGFetchForSingleItem() we queue an item onto checkpoint if the
expirytime time of the bg fetched item does not match the expiry time of the
item in the ht. But this is redundant because whenever expiry time of an item
changes we already queue it onto checkpoint.

Verified by an existing test case test_mb5215()

For records, this was originally added in the commit 9cd0a77

Change-Id: I700c4f00c3dda30fba491fcf047a86bd954d212f
Reviewed-on: http://review.couchbase.org/71623
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>