Remove nvalue in item_info
We're always using a single entry in the io vector. We've got
an idea of using multiple entries to keep xattr's in one and
the body in another. Once we decide to do that, we should
change the item_info struct to reflect that.

The old item_info was designed this way when we wanted to
support a wide range of engines we didn't have control over,
but now that we only support our own engines we don't need
that kind of flexibility (and complexity)

Change-Id: Ifeb9a2f5cd4b1beb2afa072314000bd0a663fddf
Reviewed-on: http://review.couchbase.org/71899
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>