Return ENOENT for 'warmup' stats if warmup disabled
ns_server checks the value of the 'warmup' stats group when
determining if a bucket is read - see ns_memcached/check_started. It
expects this to:

a) return 'ENOENT' - when warmup is not applicable (historically
   memcached buckets).

b) include a 'ep_warmup_thread' key - if so then it will wait until
   this has the value 'complete' before marking the bucket as ready.

So ep-engine acts as ns_server expects, if warmup is disabled (and
hence the Warmup task does not exist), return ENOENT if the 'warmup'
set of stats is requested. This means that ns_server will consider the
bucket ready and activate it.

Change-Id: I42f109146f260268f3a999e93d36e7e5c99dfada
Reviewed-on: http://review.couchbase.org/71902
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>