[Refactor]: Move queueDirty() go from KVBucket to VBucket
KVBucket::queueDirty() is a function used to put an item onto checkpoint
after it is inserted into hash table. Both hash table and checkpoint are
members of VBucket class and hence it is more appropriate to implement
this function in VBucket class.

Also, added VBUpdateReturnCtx which will indicate the KVBucket on the
actions to be taken after adding/deleting/updating an item in VBucket.

Implementing this functions in VBucket class will help in refactoring
calls to hash table and also result better abstraction.

Change-Id: I74ecdd774a88981feb482a34d375f38a12cfc900
Reviewed-on: http://review.couchbase.org/71569
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>