CheckpointTest: prepare for conversion to Typed Test
We want to expand the Checkpoint Test so it can be run using either a
(persistent) VBucket or EphemeralVBucket object, which involves
converting to a Typed Test.

Due to the implementaiton of these in GoogleTest, all member variables
of the class accessed in tests need to be prefixed with `this->` - see
https://github.com/google/googletest/blob/master/googletest/docs/AdvancedGuide.md#typed-tests

This patch does eactly that - adds `this->` to the front of all member
accesses. There's also a bunch of whitespace changes due to running
through `git clang-format`.

Change-Id: I29337a097a4f1e62ef66d847b0be34fdfb36cec5
Reviewed-on: http://review.couchbase.org/71977
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>