MB-22318: Allow for setting a full xattr
This is a temporary fix which makes us closer to feature complete
at the cost of performance.

Part of the problem here is the way we've made the xattrs work
so that we're stripping off the first part of the xpath, and
auto-creating the "container" for the xattr. We could add special
handling for the "full xattr" path spec, but then we would have
problems when people do multi-ops and their first entry isn't
a full xattr path. What we _could_ do here is to check if
we've got a case where one of the ops is a full xattr-path
and treat those separately, but if all of the xattr paths
are partial we don't have to rebuild it..

For now we'll just do the slow path to unblock functional
testing

Change-Id: Ie614efac2936b5f4438963846fe6f054e037eef2
Reviewed-on: http://review.couchbase.org/72021
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>