MB-22015: Replace bare pointers in access log
Replace the bare owning pointers used in the access log with a concrete
mutation as each mutation is owned by the access log and is at no point
moved or nulled.

The accessLog has been changed to a std::deque rather than a std::vector
as to use emplace_back on  std::deque, T only needs to be
EmplaceConstructible ratherthan the CopyConstructible required by
std::vector. This means that a move constructor is not needed for
MutationLog, which is useful as a move constructor does not quite fit
the semantics of MutationLog. Furthermore the anticiapted performance
of std::deque is not expected to differ much from std::vector.

Change-Id: I7152bac511d18679967defd47120accbdfaa10f1
Reviewed-on: http://review.couchbase.org/72037
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>