[Refactor]: Move bgFetch() from KVBucket to VBucket
BgFetch is scheduled in full eviction mode when there is a miss in hash
table lookup. As hash table is contained in VBucket it is more appropriate
to implement this function in VBucket class.

Implementing this function in VBucket class will help in refactoring
calls to hash table and also result better abstraction.

Change-Id: I117bf20b0063723fe939967c06a3933ece1cd171
Reviewed-on: http://review.couchbase.org/71472
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>