[Refactor]: Split set operation across KVBucket and VBucket.
The set(update or add new item) operation is split across KVBucket
and VBucket classes such that the HT operations involved in set are
done in Vbucket class.

This is a better abstraction and could help in modularizing HT and
in-memory operations involving sets/deletes and other operations which
involve updating/adding/deleting items.

Change-Id: Ief76f255c4676a03ad555eb6e31fde072f571d1c
Reviewed-on: http://review.couchbase.org/71543
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>