[Refactor]: Split replace operation across KVBucket and VBucket classes
The replace (set a new value to an existing value) operation is split
across KVBucket and VBucket classes such that the HT operations and other
in memory operations involved are done in VBucket class.

This is a better abstraction and could help in modularizing HT and
in-memory operations involving sets/deletes and other operations which
involve updating/adding/deleting items.

Change-Id: I76ca2e81f5478d5e925cfefce8ddd453e286319a
Reviewed-on: http://review.couchbase.org/72092
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>