[Refactor]: Split addTAPBackfillItem() operation across KVBucket and VBucket
The addTAPBackfillItem() is split across KVBucket and VBucket classes such
that the HT operations and other in memory operations involved are done
in VBucket class.

This is a better abstraction and could help in modularizing HT and
in-memory operations involving sets/deletes and other operations which
involve updating/adding/deleting items.

Change-Id: Iae563b762d67be3d374d81581a7a95779353ebc6
Reviewed-on: http://review.couchbase.org/72093
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>