[Refactor]: Split setWithMeta operation across KVBucket and VBucket classes
The setWithMeta (non front end set op) is split across KVBucket and VBucket
classes such that the HT operations, conflict resolution and other
in memory operations involved are done in VBucket class.

This is a better abstraction and could help in modularizing HT and
in-memory operations involving sets/deletes and other operations which
involve updating/adding/deleting items.

Change-Id: Id73c9eeddc930c3ae95af8d4ed9ad2d7b0d90ebd
Reviewed-on: http://review.couchbase.org/72161
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>