MB-19955: Set connectionManagerInterval to one second
Unfortuately due an error with the following merge
Ib6f810c42ed6a4a7f473d1f157e890b0593387cb the connection manager
interval was set from one to two.

The result is that even though the noop interval setting remained at
one when there is no DCP traffic we snooze for the connection manager
interval before sending a noop.  This caused noops to be sent every two
seconds as opposed to every second.

This patch changes the connection manager interval back to one second.
In addition it checks when setting the noop interval that it is a
multiple of the connection manager interval.  If not an ENGINE_EINVAL
error is returned.

A corresponding test has also been added.

Change-Id: I631f864ab88da65fd1efae71f17af79263a337ce
Reviewed-on: http://review.couchbase.org/72180
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>