mcbp_validate_response_header: Return AssertionResult
This provides a new overload of the function which returns googletest's
AssertionResult. This will better help debug the context of the
originating test case (rather than in testapp_binprot).

The new function is intended to be used with gtest's
`{EXPECT,ASSERT}_PRED_FORMAT_N`, or even as `ASSERT_TRUE`.

More specifically, it is intended to be used as a building block for
other gtest formatters.

Returning an AssertionResult rather than failing directly within the
test allows the actual failure to be reported in a custom manner (i.e.
with a custom error message), but more importantly it also allows
the file/line of the calling test to show up in the log, rather than the
file/line of the helper function.

Typically you would do something like this:

```
::testing::AssertionResult doCommandOk(const char *, Command& c) {
    auto resp = doCommand(c);
    auto rv = mcbp_validate_response_header(resp.getHeader()
        cmd.getOp(),
        PROTOCOL_BINARY_RESPONSE_SUCCESS,
        false);
    if (!rv) {
        return rv;
    }
    // Do further checks..
    return ::testing::AssertionSuccess();
}

....

EXPECT_CMD_OK(BinprotSubdocCommand(PROTOCOL_BINARY_CMD_SUBDOC_EXISTS)
                .setKey("docid")
                .setPath("nonexist_path"));
```

More concrete examples can be seen in subsequent commits :)

Change-Id: I6e4cdf5797fe0e9e97a876118d9ab5d703e92bbb
Reviewed-on: http://review.couchbase.org/72887
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>