Retain the value for datatype PROTOCOL_BINARY_DATATYPE_XATTR on expiry
On expiry of an item with a body containing value, user xattrs and
system xattrs, instead of deleting the body, the body is being
retained because mobile clients can still query for the system xattrs
of deleted/expired items. Note that this is still inefficient as the
whole body is being retained instead of just the system xattrs.

Change-Id: Ib5e55bc6efabb23f35dc8cfebb8e762755050599
Reviewed-on: http://review.couchbase.org/72256
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>