[Refactor]: Rename addTAPBackfillItem() to addBackfillItem()
Actually addTAPBackfillItem() is used for both TAP backfill and DCP
backfill. The function basically puts the items read directly onto vb queue
for persistence, rather than putting them on checkpoint.

It is good to remove TAP from the name so that we do not get confused,
especially when we are thinking of TAP deprecation and in Ephemeral buckets
where we do not support TAP or persistence.

Change-Id: Ie70660aa0d0c2e6f745227b2447bf0daf36ee565
Reviewed-on: http://review.couchbase.org/72263
Reviewed-by: Will Gardner <willg@rdner.io>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>