MB-22436: Unify NUMA enabling and logging
Previously the NUMA configuration was done in two parts - first we
attempt to set the desired NUMA policy, recoding the result in an
enum; then later once logging is enabled we switch on the result and
form a message.

There is about to be a fourth result (interleave selected but failed),
and for that we want to include the reason why it failed. To
facilitate this refactor the code so the status message is formed
directly by the configuration code (which has the context to form any
error message).

Change-Id: Ibaa0f854aad9e7e47476e2a93bd693b427e695b4
Reviewed-on: http://review.couchbase.org/72353
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>