Remove support for DCP NO_VALUE (moved to core)
As part of adding support for XATTRs over DCP we can't strip
off the value part in ep-engine. The memcached core will
snoop the DCP OPEN and ADD STREAM requests and pick up the
flag to strip off bodies.

The removed test cases tested the cases where we want to
strip off the values (which we no longer can do due to the
fact that the frontend may need them in order to pick out
the xattrs).

Note that this patch depend on the patch in memcached
which implement stripping of the value. (given that
there isn't a compile time dependency I'm not reusing
the Change-Id as jenkins don't play very well with
patches based on top of the linked patch)

Change-Id: I86ed8c2844f70d5aa1630339ae5859a731e557bd
Reviewed-on: http://review.couchbase.org/72398
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>