MB-22234: Adds statistics for MBCP response status
Adds counters to track each MCBP response status sent.

Allows access to these new stats using a stat command with responses.
It will return a JSON object where keys are the status hex and
values are the count.

A status is counted for each request even if a response is not sent back
to the requester. It's important to note that a single status is counted
meaning that multi-responses such as STATS only record a single status.

A total count of all MCBP errors (just excludes success) is now also
included in an all stats command.

Change-Id: I8ffa33f6388f8bc039ca067d1a5daa9e4bd6a632
Reviewed-on: http://review.couchbase.org/72791
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>