Push FlusherCB construction down to EPBucket subclass
makeVBucket() takes a parameter of a flusherCB - a callback object to
be invoked when 'interesting' items are enqueued onto the
CheckpointManager. This callback is only needed for persistent (EP)
bucket types, for Ephemeral it is not necessary.

As such, push the construction of this callback to down into
EPBucket's implementation, removing it from makeVBucket()'s signature.

Change-Id: Ia657bb2e097f5b274176809095adcecc1eae4fa4
Reviewed-on: http://review.couchbase.org/72801
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>