MB-22485: Restore datatype correctly when reading from disk
We do store the datatype as part of storing the object,
but when we read the object back in we did not pick out
the correct value (we tried to determine all over again,
which didn't account for XATTRs etc).

We do have traces of the old datatype support we had
in here, and I have my doubts that it all fit together
in the big picture. Before we allow that to be enabled
we _have_ to do a full review of the current pieces
all over again.

Change-Id: Ib73a1bc20d49f13eddfa514e4f43fd5a801c080c
Reviewed-on: http://review.couchbase.org/72790
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>