Remove Flusher statistics from Buckets without a Flusher
There are a number of flusher-related statistics which are only
applicable to buckets which have a Flusher (aka Ephemeral buckets). As
such they shouldn't be returned when the underlying kvBucket doesn't
have a Flusher.

To minimise the exposure of the different KVBucket subclasses to
EvenutallyPersistentEngine, use the presence (or absence) of a Flusher
on the underlying kvBucket to control if we print statistics for it or
not.

Change-Id: I038a6cbcbca8784c5809c0fb8547ca01c20889f8
Reviewed-on: http://review.couchbase.org/72808
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>