MB-22316: Use the same hashing for plain password as ns_server
Use the same password hashing as ns_server for plain text
passwords. This allows for ns_server to add the "Administrator"
role to our config files without having to force the user
to change the password first. Please note that until the
administrator change the password it is only possible to
log in using PLAIN authentication (as the user won't have
any of the SCRAM credentials available).

The cbsasl_pwfile_test is removed as it is no longer
valid (and the same functionality is added in cbsasl_pwconv_test)

Change-Id: I12e22a6dd9cfa479800e849872d62f2b87fb1061
Reviewed-on: http://review.couchbase.org/73004
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>