Reapply "MB-22524: Add XATTRs for DCP Deletion & Expiration messages"
This reverts commit d81776fb6d3ac3d67181ec61209db237a8224bae.

The code is changed to use IO vectors for dcp_delete (in order to
allow the xattrs to be passed on without having to reallocate a
buffer) and the length field of the first io vector had an incorrect
length (added the key twice).

I've filed MB-22601 to build a full DCP test suite for memcached and
ep-engine to detect such regressions earlier.

Change-Id: I7ea4e6bbca1ca293c299db3637a49e3605f8181f
Reviewed-on: http://review.couchbase.org/73048
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>