Refactor: move cbsasl stuff from Couchbase:: to cb::sasl::
cbsasl started off as isasl as we had to create our own sasl
implementation due to the fact that cyrus sasl didn't work
on Windows. At some point we rebranded it to cbsasl and
changed the prefix to cbsasl instead of sasl to all of the
methods to avoid conflicts if one (for some reason) linked
with cyrus sasl (one of the dependencies could have been
prelinked with that).

It's been a long time since we could replace cbsasl with
cyrus sasl if we wanted to (in case it provided additional
mechs), and I think it is time to make it a C++ library
and easier to use from _our_ code base rather than trying
to stick with the function naming from the SASL library.

As a start we can put the internal classes we've got
in cbsasl into cb::sasl instead of the old and deprecated
Couchbase namespace.

Change-Id: Ic8461e53b0d6c0fee92bffdc4a5cc018550d7c9e
Reviewed-on: http://review.couchbase.org/73067
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>