[Refactor]: Do not pass HT lock as a param to VBucket::queueDirty()
VBucket::queueDirty() must just put an item on a queue for persistence
and/or replication. Hence the HT lock and notification logic are removed
from it. It should be upto the VBucket to decide when to release the HT
lock and when to notify any clients of a new seqno

Change-Id: I8e72d74d32190953056609f5991e469b6b909509
Reviewed-on: http://review.couchbase.org/73038
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>