MB-16181: Collections on/off config option
Whilst collections exists in an incomplete and "prototype" state we
need to have a way to enable the feature for development.

This config switch should be viewed as temporary whilst collections is
not complete or production ready.

The first piece of 'dangerous' collections related functionality to be
enabled by this switch is the persistence of the document namespace.

Persisting the document namespace allows collections to store system
events, default-collection document and collection-documents, but will
break upgrades because existing data has no namespace.

Change-Id: Ic9ab39fdf572f4feebd497135e599b90ffc1221b
Reviewed-on: http://review.couchbase.org/73122
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>