MB-22623: Fix parsing of stats subdoc_execution
We switched the implementation to use our internal client to
support all of the various ways to communicate with the server
(IPv4/6, SSL, SASL mechanisms etc).

We had a special implementation for the normal "get timings"
commands, but we use "stats" in order to fetch the subdoc_execution
timings and our client returned a JSON payload where it
automatically created a "KV" pair from each response as a
stat call could return multiple values. The old code expected
just a single return and handled that, and that logic got
lost in the refactor to the new client.

Change-Id: I07b305697c4718761cd81226871ad1febd0bc8c1
Reviewed-on: http://review.couchbase.org/73155
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>