MB-22041 [2/N]: Change existing tests to accommodate dynamic threads
Previously, the number of threads (Reader, Writer, NonIO, AuxIO) allowed
to run concurrently could be tuned down, with excess threads looping on
sleeping and checking for work. Crucially, they were still able to clean
up dead tasks.

With MB-22041 changes allowing one to dynamically stop and start
threads, some tests would fail as they set some thread types to 0.

When attempting to stop a task group, all the tasks in the group are
cancelled. If there are 0 threads runnning for that task type, these
tasks cannot be cleaned up, and _stopTaskGroup loops endlessly as they
remain in the queues.

Setting the threads back up to 1 resolves this.

Change-Id: I2ef9618fe3b2389aa6c995d8333a813137d68411
Reviewed-on: http://review.couchbase.org/73303
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>