[Refactor]: Move setMaxCasAndTrackDrift() in a func that calls queueDirty()
setMaxCasAndTrackDrift() is called in setWithMeta and deleteWithMeta.
In the both the place it just precedes queueDirty(), and also has a
strict requirement that it must be just called before queueDirty().

Hence moving it to a separate function that calls queueDirty() immediately
after tracking cas drift is better for maintainance and also helps to keep
track of things if we want to move around queueDirty() calls.

Change-Id: Ie0c42f05fba07bde7c26cdec123daae62bc99725
Reviewed-on: http://review.couchbase.org/73141
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>