MB-22425: Don't acquire Configuration mutex in itemAllocate()
In EPEngine::itemAllocate we need to check if priv_nbytes exceeds the
configured limit. This was done by calling
configuration.getMaxItemPrivilegedBytes() in the method, which
requires taking Configuration::mute and searching the map of
attributes for the given param.

This is costly, and resulted in a ~8% drop in throughput in the 256B
50/50 pillowfight test.

To address this, add a new maxItemPrivilegedBytes member variable to
EPEngine which is initialized at object construction from
configuration.getMaxItemPrivilegedBytes(). To handle dynamic changes
of the config setting, also register a ValueChangedListener which
updates EPEngine::maxItemPrivilegedBytes when required.

Local testing (mancouch) suggests this resolves the performance
regression.

Change-Id: I957ffbc8f78b661c0d6cdab90d7d43dc0d6bdc00
Reviewed-on: http://review.couchbase.org/73307
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>