Serialize item add/update/softDelete at EphemeralVBucket level
In an EphemeralVBucket we need to store items sequentially apart
from storing it in HashTable. We intend to store these items in a
data structure, a 'sequence list', which may be implemented as a
linked list or a skiplist. We must serialize the adds/updates/softDeletes
on the sequence list.

Currently seqno is generated in the CheckpointManager and is synchronized
on 'queueLock' in the CheckpointManager. CheckpointManager relies on seqno
for its meta(dummy) items and also self generates them (that is generates
seqno when no external item is added to it). It is therefore not a trivial
task to move the seqno generation entirely to the EphemeralVBucket.

This commit introduces a new lock at the EphemeralVBucket level 'orderLock'.
Using this lock ensures that at the VBucket level we have correct ordering
of items even though seqno is generated by the CheckpointManager later.
The seqno generated by the CheckpointManager later is updated in the
sequence list.

All operations/data structures that rely on ordered sequence of items
must grab i) orderLock in 'EphemeralVBucket' and then ii) queueLock in
'CheckpointManager'.

The commit also derives child class (EphemeralVBucket) functions for
add/update/softDelete and gives an overview of what changes will be
made in EphemeralVBucket when sequence list is added pretty soon.

Pros of this approach:
1. CheckpointManager logic is unchanged and hence smaller chance of
   regression.
2. Difference in operation of VBucket vs EphemeralVBucket level is little.

Cons of this approach:
1. Some rogue functions not grabbing 2 locks (orderLock and queueLock)
   correctly can screw up things. But I think since we have very few
   functions handling add/update/softDelete, it should not be a big problem.
2. There is a tiny (very tiny :) ) window where we the item in the
   'sequence list' though is in correct order, does not have a seqno.
   Hence special care must be taking during DCP backfill from the
   sequence list.

Change-Id: I99e928c75f9eee101bba8852b6cc531bb8353886
Reviewed-on: http://review.couchbase.org/73275
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>