MB-22041 [3/N]: Allow dynamic changes to number of threads
Reader, Writer, AuxIO and NonIO threads can now be dynamically stopped
and started using the interface previously used to set the maximum
number of these threads - i.e., setMaxReaders.

The behaviour is outwardly the same - the max controls how many tasks of
a given type can be run concurrently, but this can now be dynamically
tuned both upwards and downwards.

Change-Id: I8e10487e5b57466fbe299e7e043a90dd6c8d5aa8
Reviewed-on: http://review.couchbase.org/73304
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>