EPStats: Use RelaxedAtomic for counter-style statistics
The majority of the statistics in EPStats are simple counters which
are updated by multiple threads, and hence are implemented using
std::atomic. std::atomic uses sequently consistent memory ordering by
default, which is overly strict - we do not need ordering or
synchronization from these variables, just atomic updates.

As such change to implement in terms of RelaxedAtomic. This results in
a small (<2%) improvement on op/s using the following pillowfight
benchmark on mancouch (2 nodes):

    cbc-pillowfight -U couchbase://127.0.0.1:12000/default --min-size 256 --max-size 1024 -I 5000000 -B 1000 -t 12 -r 50

Change-Id: Ie241eec8f0c8282eb958f468dde962ca1c79bf84
Reviewed-on: http://review.couchbase.org/73313
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>