MB-16181: Use cb::const_char_buffer over const std::string ref
It will be easier/cheaper to integrate the DCP side of the collections
work if we can update a collection from a DCP message. The "easy/cheaper"
benefit being that we can just use a cb::const_char_buffer initialised to
the packet's collection name and length data, rather than being forced to
create and copy into a std::string.

Change-Id: I58858e9a1f4fe7f0c6b382160764f68c63a6fb97
Reviewed-on: http://review.couchbase.org/73444
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>