MB-19339: Refactor the way we access RBAC
Earlier we grabbed a shared pointer to the RBAC database with the
idea that the shared_ptr magic should ensure that the RBAC database
would be alive until the thread using it would complete. Unfortuntately
this wouldn't fly, so I've refactored the system to work without the
shared pointers.

The most performance sensitive method in the system is
PrivilegeContext::check() as it is called for almost all operations.
It is implemented by checking if the context is created from the
same revision of the privilege database as we're currently using,
and if that is true we use the internal bitmask.

In order to create the PrivilegeContext (which happens every time a
user connects, authenticates or switch bucket) we use a RW lock.
For now this means that we _COULD_ starve the writer and delay
the update of the privilege context, but I'm going to stick my
head in the sand for now and ignore that problem until we've gotten
the rest of the system in place. We'll grab a read lock to the
privilege database, look up the user and release the lock.

Change-Id: I21e58ceb5e084549cbcce939c7ef2c41530bdc99
Reviewed-on: http://review.couchbase.org/73487
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>