[Refactor]: Remove force param from KVBucket::deleteItem and VBucket::deleteItem
Force param in these functions is always passed as false. Also the code path
related to force == true leads to a possible exception (never executed and
hence not seen so far). It is better to get rid of this unexecuted code path.

Further, I see that force == true is unused atleast from Watson.

Change-Id: I184dd14a8fc58fbb89dd98e170414f3504711315
Reviewed-on: http://review.couchbase.org/73477
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>