Pass stream_t by ref, not by copy for some DCP paths
Spotted whilst single stepping that some paths are copying the
shared pointer down the call chain, increasing the ref count for
each new copy.

In some cases we would abort, requiring an unwind of all the
increments.

So pass by ref and only take a copy when we really need to extend
the lifetime of the object.

Change-Id: Ic03087ad1a6d06d4792472f57271f80185fa0c71
Reviewed-on: http://review.couchbase.org/73620
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>