MB-19339: Add a privilege debug flag
This allows for a single place to enable/disable this. The
intention is that when we're running in privilege debug mode
we'll allow the action even if the privilege check say no.

A user may then start the server and log in as an unprivileged
user and run his app, and then look in the logs to see all
privileges he would need to have and build up the correct
(minimum) profile.

This patch only includes the framework (as the privilege
checks isn't added yet).

Change-Id: I043de5c5b1c02b7c696abce696b72858f9238b27
Reviewed-on: http://review.couchbase.org/73693
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>