MB-22822: Reduce allocation of Blob by 2 bytes
We were allocating the sizeof(Blob) + space needed to store the
data. However because we had a 1 byte array (data) in order to access
the stored data we were wasting 1 byte by allocating 1 byte too many.

Additionally, because the largest constituent data member is a uint32_t
the compiler would pad a Blob with 1 byte in order to make the
sizeof(Blob) 12. We can account for this by having a 2 byte array
instead and, again, take this in to account when allocating memory so we
do not waste the 2 bytes.

Change-Id: Id0b7203b75bed2de040f6841e8c1a6c16ce35f96
Reviewed-on: http://review.couchbase.org/73739
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>