Add a HashTable function to remove (not delete) a StoredValue from it
In Ephemeral Buckets we have a requirement to remove items from HashTable
and mark them as stale rather than deleting them immediately.

Here we create a general HashTable function to just remove a StoredValue
from it and not delete it. Also since this logic is pretty much the
same as HashTable::unlocked_del() we reuse the code by calling remove
before deletion.

Change-Id: Id84f75a2c4648f58727eff9099c99259e86c565e
Reviewed-on: http://review.couchbase.org/73278
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>