Always enforce 'VALUE_ONLY' eviction policy in EphemeralBuckets
In EphemeralBuckets we want only logic related to VALUE_ONLY eviction
policy. That is because a key not present in HashTable implies key not
present at all.

This is needed because we do not have proper abstraction for
VALUE_ONLY and FULL_EVICTION policies. Both these do not make much
sense from Ephemeral Buckets, where eviction would mean deletion.
In Ephemeral buckets we would have desired to have NO_EVICTION
policy which would be very close to VALUE_ONLY policy because of
"a key not present in HashTable implies key not present at all".

Now instead of rewriting many functions that use myriad of if-else
for Ephemeral Buckets, we can use those functions directly in
Ephemeral Buckets with VALUE_ONLY policy enforced.

Note: This should not be confused with the eviction algorithm
      that we are going to use like NRU, FIFO etc.

Change-Id: Ied0dbe615dff8fb1625b96a355efeba2117747b1
Reviewed-on: http://review.couchbase.org/73864
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>