MB-22513: Only calculate hash bucket once
In codepaths that call HashTable::unlocked_addNewStoredValue, we often
find the hash bucket at some point earlier in the path and then again in
HashTable::unlocked_addNewStoredValue. Therefore we end up calculating
the hash twice.

We change HashTable::unlocked_addNewStoredValue to not calculate the
hash bucket and instead have it passed in as an argument. This requires
some changes earlier in the call path which is what the bulk of this
commit is.

Change-Id: Ib0c630030233b35f11336b8d7cf181370d17f146
Reviewed-on: http://review.couchbase.org/73953
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>