Make VBucket abstract class. Define virtual funcs in derived classes
Virtual functions updateStoredValue(), addNewStoredValue() and
softDeleteStoredValue() are implemented in EPVBucket and EphemeralVBucket
classes. They are made abstract functions in base class VBucket.

Due to this, an object of the abstract class VBucket cannot be created
anymore. So in unit test code we should create objects of EPVBucket and
EphemeralVBucket and run tests on them.

This commit, in unit testing, creates only the objects of EPVBucket and
runs tests on them. EphemeralVBucket to follow soon.

Change-Id: I09c9ef4d903729fd6c6569008ee30697342706fb
Reviewed-on: http://review.couchbase.org/73999
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>